+++ execution_engine/src/core/resolvers/v1_function_index.rs
@@ -57,7 +57,7 @@ pub(crate) enum FunctionIndex {
     DictionaryPutFuncIndex,
     LoadCallStack,
     LoadAuthorizationKeys,
-    NextAddress,
+    RandomBytes,
 }
 
 impl From<FunctionIndex> for usize {